Refactor 03_02 according to #61

+++ examples/chapter03_02/src/mcal/avr/mcal_gpt.h
@@ -25,7 +25,7 @@
     namespace gpt
     {
       typedef void          config_type;
-      typedef std::uint64_t value_type;
+      typedef std::uint32_t value_type;
 
       void init(const config_type*);
